Update app.py
Made the loading animation appear on the first question (previously only appeared on second and later questions). Added a warmup when space is first opened, speeds up the speed tremendously (there is a timer in the bottom corner, tracking the response time to each query. Previously, asking what the capital of germany was took 25 seconds. Now it takes 1.5).